You can now customize sprite head color!
Have successfully come up with a method of customizing a single pice of a sprite and implimented it. You can select the color of the head of the test sprite on the login screen, and that color will now be displayed in game when you are moving around!

Have not yet implimented this for remote players. It only works when viewing yourself at the moment.